+++ docs/changelog.rst
@@ -6,6 +6,11 @@ Changelog
 `Unreleased`_
 -------------
 
+Added
+~~~~~
+
+- ``--verbosity`` CLI option to minimize the error output. `#598`_
+
 `1.8.0`_ - 2020-06-15
 ---------------------
 
@@ -1118,6 +1123,7 @@ Fixed
 .. _#612: https://github.com/kiwicom/schemathesis/issues/612
 .. _#600: https://github.com/kiwicom/schemathesis/issues/600
 .. _#599: https://github.com/kiwicom/schemathesis/issues/599
+.. _#598: https://github.com/kiwicom/schemathesis/issues/598
 .. _#596: https://github.com/kiwicom/schemathesis/issues/596
 .. _#594: https://github.com/kiwicom/schemathesis/issues/594
 .. _#586: https://github.com/kiwicom/schemathesis/issues/586